refactor(driver): organize driver tests
