+++ compio-driver/tests/file.rs
@@ -1,10 +1,14 @@
-use std::{io, time::Duration};
+use std::io;
 
 use compio_buf::{arrayvec::ArrayVec, BufResult};
 use compio_driver::{
-    op::{Asyncify, CloseFile, OpenFile, ReadAt},
-    OpCode, Proactor, PushEntry, RawFd,
+    op::{CloseFile, OpenFile, ReadAt},
+    Proactor, PushEntry, RawFd,
 };
+
+mod utils;
+
+use utils::push_and_wait;
 
 #[cfg(windows)]
 fn open_file_op() -> OpenFile {
@@ -41,20 +45,6 @@ fn open_file_op() -> OpenFile {
     OpenFile::new(CString::new("Cargo.toml").unwrap(), flags, 0o666)
 }
 
-fn push_and_wait<O: OpCode + 'static>(driver: &mut Proactor, op: O) -> (usize, O) {
-    match driver.push(op) {
-        PushEntry::Ready(res) => res.unwrap(),
-        PushEntry::Pending(user_data) => {
-            let mut entries = ArrayVec::<usize, 1>::new();
-            while entries.is_empty() {
-                driver.poll(None, &mut entries).unwrap();
-            }
-            assert_eq!(entries[0], *user_data);
-            driver.pop(user_data).unwrap()
-        }
-    }
-}
-
 #[test]
 fn cancel_before_poll() {
     let mut driver = Proactor::new().unwrap();
@@ -81,17 +71,6 @@ fn cancel_before_poll() {
 
     let op = CloseFile::new(fd);
     push_and_wait(&mut driver, op);
-}
-
-#[test]
-fn timeout() {
-    let mut driver = Proactor::new().unwrap();
-
-    let mut entries = ArrayVec::<usize, 1>::new();
-    let err = driver
-        .poll(Some(Duration::from_secs(1)), &mut entries)
-        .unwrap_err();
-    assert_eq!(err.kind(), io::ErrorKind::TimedOut);
 }
 
 #[test]
@@ -123,30 +102,4 @@ fn register_multiple() {
 
     let op = CloseFile::new(fd);
     push_and_wait(&mut driver, op);
-}
-
-#[test]
-fn notify() {
-    let mut driver = Proactor::new().unwrap();
-
-    let handle = driver.handle().unwrap();
-
-    let thread = std::thread::spawn(move || {
-        std::thread::sleep(Duration::from_secs(1));
-        handle.notify().unwrap()
-    });
-
-    let mut entries = ArrayVec::<usize, 1>::new();
-    driver.poll(None, &mut entries).unwrap();
-
-    thread.join().unwrap();
-}
-
-#[test]
-fn asyncify() {
-    let mut driver = Proactor::new().unwrap();
-
-    let op = Asyncify::new(|| BufResult(Ok(114514), ()));
-    let (res, _) = push_and_wait(&mut driver, op);
-    assert_eq!(res, 114514);
 }
